This flag tells npm to ignore peer dependency conflicts and install packages anyway, which should allow your build to complete successfully. This is a common approach when working with packages that haven't yet updated their peer dependency requirements for newer versions of React.
This is a quick fix to get your GitHub Actions workflow running. For a more permanent solution, you might consider:

Downgrading React to version 18.x to match the requirements of tech-stack-icons
Looking for an updated version of tech-stack-icons that supports React 19
Finding an alternative to tech-stack-icons that's compatible with React 19
Would you like me to help with any of these alternative approaches as well?